# shared/openapi/api.yaml
...
openapi: 3.0.0
info:
  title: Montgomery Guardian API
  version: 1.0.0
  description: Backend API for Montgomery Guardian platform
servers:
  - url: https://montgomery-guardian-api.replit.app
    description: Production (Replit)
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/vision/analyze:
    post:
      summary: Analyze photo for 311 report
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                lat: { type: number }
                lng: { type: number }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VisionAnalysisResult' }
components:
  schemas:
...
    VisionAnalysisResult:
      type: object
      properties:
        incidentType: { type: string }
        severity: { type: string }
        confidence: { type: number }
        description: { type: string }
        prefilledForm: { type: object }
...